Significant rework of all the networking code.

* Removes dead/redundant code
* Simplifies network flow
* Removes string replacements
* _Starts_ standardizing string formatting method (still a lot more to do here)
* Removes redundant calls to DataManager

~There's still some cleanup that needs to happen between going to `API.get()` directly vs going through `DataManager.get_content()`~

Added benefit of items loading faster now.  On one computer from 6 seconds to 4 seconds.  On the other from 2.5 seconds to 2 seconds flat.